+++ libatomic.cc
@@ -13,7 +13,7 @@ int atomic_load_explicit(struct atomic_object *obj, memory_order order)
 {
 	DBG();
 	model->switch_to_master(new ModelAction(ATOMIC_READ, order, obj));
-	return thread_current()->get_return_value();
+	return (int) thread_current()->get_return_value();
 }
 
 void atomic_init(struct atomic_object *obj, int value)
